test_fabrique/requirements.txt
Django==2.2.10
django-filter==21.1
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
